requirements.txt
ruamel.yaml==0.15.35
boto3==1.6.5
docker==3.1.1
jinja2==2.10
